Move changes grouping by codeowner code to dedicated package
