+++ src/main/kotlin/com/github/fantom/codeowners/grouping/changes/SetCodeownersChangesGroupingAction.kt
@@ -1,0 +1,16 @@
+package com.github.fantom.codeowners.grouping.changes
+
+import com.github.fantom.codeowners.CodeownersManager
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.components.service
+import com.intellij.openapi.vcs.changes.actions.SetChangesGroupingAction
+
+class SetCodeownersChangesGroupingAction : SetChangesGroupingAction() {
+    override val groupingKey: String get() = "codeowners"
+
+    override fun update(e: AnActionEvent) {
+        super.update(e)
+        val manager = e.project?.service<CodeownersManager>()
+        e.presentation.isEnabledAndVisible = e.presentation.isEnabledAndVisible && manager?.isAvailable ?: false
+    }
+}
